feat: ex01 operation can handle doubles, overflow is handled

+++ ex01/RPN.hpp
@@ -6,7 +6,7 @@
 /*   By: zpalotas <zpalotas@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/03/13 17:32:45 by zpalotas          #+#    #+#             */
-/*   Updated: 2026/03/13 19:52:23 by zpalotas         ###   ########.fr       */
+/*   Updated: 2026/03/20 14:30:32 by zpalotas         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,13 +18,14 @@
 # include <sstream>		//stringstream
 # include <stack>
 # include <algorithm>
+# include <climits>
 # include "colors.hpp"
 
 class RPN
 {
 	private:
 		RPN();
-		std::stack<int> expr;
+		std::stack<double> expr;
 
 		void addition();
 		void substraction();
